+++ pathdiscov/run_standard.pl
@@ -76,7 +76,7 @@ GetOptions (	'outputdir=s' => \$outputdir,
 $strstages = join(' ', @stages);
 # Set default stages if not set by options
 if( $strstages eq "" ) {
-    $strstages = 'step1 host_map quality_filter ray2_assembly iterative_blast_phylo orf_filter';
+    $strstages = 'step1 host_map quality_filter ray2_assembly iterative_blast_phylo';
 }
 
 print("Stages to be run: $strstages\n");
@@ -131,10 +131,6 @@ else
 
 print("\n-|-------unassembled reads-------|-\n");
 print_system("$path_scripts/pathogen.pl --sample $sample --command iterative_blast_phylo_2 --paramfile $paramFile --outputdir $outputdir --R1 $r1 --R2 $r2");
-
-print("\n-|-------blastx/diamond contigs-------|-\n");
-$r1 = "$outputdir/orf_filter/orf_filter.contig";
-print_system("$path_scripts/pathogen.pl --sample $sample --command iterative_blast_phylo_3 --paramfile $paramFile --outputdir $outputdir --R1 $r1 --fastafile yes");
 
 print("\n-|-------read counts-------|-\n");
 print_system("mkdir -p $outputdir/output");